backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
openai>=1.60.0
httpx>=0.27.0
